Vision/AutoMLVisionClassifier/requirements.txt
fastapi==0.61.1
python-dotenv==0.14.0
uvicorn==0.12.1
ObjDict==0.4.4
azureml-core==1.22.0
azureml-dataset-runtime==1.22.0
azureml-contrib-dataset==1.22.0
azureml-telemetry==1.22.0
azureml-automl-core==1.22.0
azureml-automl-runtime==1.22.0
azureml-train-automl-client==1.22.0
azureml-defaults==1.22.0
azureml-interpret==1.22.0
azureml-train-automl-runtime==1.22.0
azureml-train-automl==1.22.0
torch==1.7.1
torchvision==0.8.2


